fake-reviews-app/pyproject.toml
[tool.poetry]
name = "fake-reviews-app"
version = "0.1.0"
description = ""
authors = ["jvk <jvkkoppineedi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
torch = "^2.3.0"
torchvision = "^0.18.0"
torchaudio = "^2.3.0"
nltk = "^3.8.1"
transformers = "^4.41.2"
joblib = "^1.4.2"
numpy = "^1.26.4"
scikit-learn = "^1.5.0"
xgboost = "^2.0.3"
flask-cors = "^4.0.1"
pyspellchecker = "^0.8.1"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
